# Conflicts:
#	buildconfig/appcompile/compile-libraries.gradle
#	features/content/createpost/src/main/java/com/tokopedia/createpost/di/CreatePostModule.kt
#	features/discovery/autocomplete/src/main/java/com/tokopedia/autocomplete/AutoCompleteActivity.java
#	features/merchant/shop_settings/src/main/java/com/tokopedia/shop/settings/notes/view/fragment/ShopSettingsNotesListFragment.kt

+++ features/merchant/product_manage_common/src/main/java/com/tokopedia/product/manage/common/feature/draft/data/model/shipment/ShipmentInputModel.kt
@@ -1,7 +1,7 @@
 package com.tokopedia.product.manage.common.feature.draft.data.model.shipment
 
 import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
+import kotlinx.parcelize.Parcelize
 
 /**
  * Created by faisalramd on 2020-03-22.
@@ -17,5 +17,5 @@ data class ShipmentInputModel (
 
 @Parcelize
 data class CPLModel (
-        var shipmentServicesIds: ArrayList<Int>? = arrayListOf()
+        var shipmentServicesIds: ArrayList<Long>? = arrayListOf()
 ): Parcelable
